+++ src/array/bounded-length-tuple.test.mts
@@ -10,27 +10,27 @@ import {
   type ValidationError,
   validationErrorsToMessages,
 } from '../utils/index.mjs';
-import { arrayBoundedLength } from './array-bounded-length.mjs';
+import { boundedLengthTuple } from './bounded-length-tuple.mjs';
 
-describe(arrayBoundedLength, () => {
+describe(boundedLengthTuple, () => {
   describe('arg patterns', () => {
     test('without explicit default value', () => {
       assert.deepStrictEqual(
-        arrayBoundedLength(2, 4, number()).defaultValue,
+        boundedLengthTuple(2, 4, number()).defaultValue,
         [0, 0],
       );
     });
 
     test('with explicit element default value', () => {
       assert.deepStrictEqual(
-        arrayBoundedLength(2, 4, number(5)).defaultValue,
+        boundedLengthTuple(2, 4, number(5)).defaultValue,
         [5, 5],
       );
     });
 
     test('with explicit default value override', () => {
       assert.deepStrictEqual(
-        arrayBoundedLength(2, 4, number(), {
+        boundedLengthTuple(2, 4, number(), {
           typeName: 'ys',
           defaultValue: [1, 2, 3],
         }).defaultValue,
@@ -39,7 +39,7 @@ describe(arrayBoundedLength, () => {
     });
   });
 
-  const xs = arrayBoundedLength(2, 4, number(), {
+  const xs = boundedLengthTuple(2, 4, number(), {
     typeName: 'xs',
     defaultValue: [1, 2, 3],
   });
@@ -260,17 +260,17 @@ describe(arrayBoundedLength, () => {
     const dynamic: number = 3;
 
     // min ∈ SmallUint & max ∉ SmallUint -> MinLengthTuple<min, A>
-    const atLeast = arrayBoundedLength(2, dynamic, number());
+    const atLeast = boundedLengthTuple(2, dynamic, number());
 
     expectType<TypeOf<typeof atLeast>, MinLengthTuple<2, number>>('=');
 
     // min ∉ SmallUint & max ∈ SmallUint -> MaxLengthTuple<max, A>
-    const atMost = arrayBoundedLength(dynamic, 5, number());
+    const atMost = boundedLengthTuple(dynamic, 5, number());
 
     expectType<TypeOf<typeof atMost>, MaxLengthTuple<5, number>>('=');
 
     // min ∉ SmallUint & max ∉ SmallUint -> readonly A[]
-    const unbounded = arrayBoundedLength(dynamic, dynamic, number());
+    const unbounded = boundedLengthTuple(dynamic, dynamic, number());
 
     expectType<TypeOf<typeof unbounded>, readonly number[]>('=');
