+++ src/array/index.mts
@@ -1,8 +1,8 @@
-export * from './array-at-least-length.mjs';
-export * from './array-at-most-length.mjs';
-export * from './array-bounded-length.mjs';
-export * from './array-of-length.mjs';
 export * from './array.mjs';
+export * from './bounded-length-tuple.mjs';
+export * from './fixed-length-tuple.mjs';
 export * from './length-constrained-array.mjs';
+export * from './max-length-tuple.mjs';
+export * from './min-length-tuple.mjs';
 export * from './non-empty-array.mjs';
 export * from './tuple.mjs';
