+++ src/array/min-length-tuple.test.mts
@@ -1,44 +1,51 @@
 import { expectType, Result } from 'ts-data-forge';
-import { type MaxLengthTuple } from 'ts-type-forge';
+import { type MinLengthTuple } from 'ts-type-forge';
 import { number } from '../primitives/index.mjs';
 import { type TypeOf } from '../type.mjs';
 import {
   type ValidationError,
   validationErrorsToMessages,
 } from '../utils/index.mjs';
-import { arrayAtMostLength } from './array-at-most-length.mjs';
+import { minLengthTuple } from './min-length-tuple.mjs';
 
-describe(arrayAtMostLength, () => {
+describe(minLengthTuple, () => {
   describe('arg patterns', () => {
     test('without explicit default value', () => {
-      assert.deepStrictEqual(arrayAtMostLength(3, number()).defaultValue, []);
+      assert.deepStrictEqual(
+        minLengthTuple(3, number()).defaultValue,
+        [0, 0, 0],
+      );
+    });
+
+    test('with explicit element default value', () => {
+      assert.deepStrictEqual(minLengthTuple(2, number(5)).defaultValue, [5, 5]);
     });
 
     test('with explicit default value override', () => {
       assert.deepStrictEqual(
-        arrayAtMostLength(3, number(), {
+        minLengthTuple(3, number(), {
           typeName: 'ys',
-          defaultValue: [1, 2, 3],
+          defaultValue: [1, 2, 3, 4],
         }).defaultValue,
-        [1, 2, 3],
+        [1, 2, 3, 4],
       );
     });
   });
 
-  const xs = arrayAtMostLength(3, number(), {
+  const xs = minLengthTuple(3, number(), {
     typeName: 'xs',
     defaultValue: [1, 2, 3],
   });
 
   type Xs = TypeOf<typeof xs>;
 
-  expectType<Xs, MaxLengthTuple<3, number>>('=');
+  expectType<Xs, MinLengthTuple<3, number>>('=');
 
   expectType<typeof xs.defaultValue, Xs>('=');
 
   describe('is', () => {
-    test('truthy case (max length)', () => {
-      const ys: unknown = [4, 5, 6] as const;
+    test('truthy case', () => {
+      const ys: unknown = [4, 5, 6, 7] as const;
 
       if (xs.is(ys)) {
         expectType<typeof ys, Xs>('=');
@@ -49,26 +56,38 @@ describe(arrayAtMostLength, () => {
       assert.isTrue(xs.is(ys));
     });
 
-    test('truthy case (empty)', () => {
-      const ys: unknown = [] as const;
+    test('falsy case 1', () => {
+      const ys: unknown = [1, 2] as const;
 
-      assert.isTrue(xs.is(ys));
-    });
-
-    test('falsy case 1 (too long)', () => {
-      const ys: unknown = [1, 2, 3, 4] as const;
+      if (xs.is(ys)) {
+        expectType<typeof ys, Xs>('=');
+      } else {
+        expectType<typeof ys, unknown>('=');
+      }
 
       assert.isFalse(xs.is(ys));
     });
 
-    test('falsy case 2 (wrong element type)', () => {
+    test('falsy case 2', () => {
       const ys: unknown = [1, '2', 3] as const;
 
+      if (xs.is(ys)) {
+        expectType<typeof ys, Xs>('=');
+      } else {
+        expectType<typeof ys, unknown>('=');
+      }
+
       assert.isFalse(xs.is(ys));
     });
 
-    test('falsy case 3 (not an array)', () => {
+    test('falsy case 3', () => {
       const ys: unknown = 'foo';
+
+      if (xs.is(ys)) {
+        expectType<typeof ys, Xs>('=');
+      } else {
+        expectType<typeof ys, unknown>('=');
+      }
 
       assert.isFalse(xs.is(ys));
     });
@@ -76,7 +95,7 @@ describe(arrayAtMostLength, () => {
 
   describe('validate', () => {
     test('truthy case', () => {
-      const ys: unknown = [4, 5] as const;
+      const ys: unknown = [4, 5, 6, 7] as const;
 
       const result = xs.validate(ys);
 
@@ -86,7 +105,7 @@ describe(arrayAtMostLength, () => {
 
       const resultValue = Result.unwrapThrow(result);
 
-      assert.deepStrictEqual(resultValue, [4, 5]);
+      assert.deepStrictEqual(resultValue, [4, 5, 6, 7]);
     });
 
     test('validate returns input as-is for OK cases', () => {
@@ -101,7 +120,7 @@ describe(arrayAtMostLength, () => {
       expect(resultValue1).toBe(input); // ✅ same reference
     });
 
-    test('falsy case 1 (not an array)', () => {
+    test('falsy case 1', () => {
       const ys: unknown = 'foo';
 
       const result = xs.validate(ys);
@@ -125,8 +144,8 @@ describe(arrayAtMostLength, () => {
       ]);
     });
 
-    test('falsy case 2 (too long)', () => {
-      const ys: unknown = [1, 2, 3, 4] as const;
+    test('falsy case 2', () => {
+      const ys: unknown = [1, 2] as const;
 
       const result = xs.validate(ys);
 
@@ -141,19 +160,19 @@ describe(arrayAtMostLength, () => {
           expectedType: 'xs',
           typeName: 'xs',
           details: {
-            kind: 'array-max-length',
-            maxLength: 3,
-            actualLength: 4,
+            kind: 'array-min-length',
+            minLength: 3,
+            actualLength: 2,
           },
         },
       ]);
 
       assert.deepStrictEqual(validationErrorsToMessages(resultError), [
-        'Error: expected array of length 3 or less but length 4 was passed.',
+        'Error: expected array of length 3 or more but length 2 was passed.',
       ]);
     });
 
-    test('falsy case 3 (wrong element type)', () => {
+    test('falsy case 3', () => {
       const ys: unknown = [1, '2', 3] as const;
 
       const result = xs.validate(ys);
@@ -179,22 +198,16 @@ describe(arrayAtMostLength, () => {
   });
 
   describe('fill', () => {
-    test('keeps numeric entries and trims extras to max', () => {
+    test('keeps numeric entries and trims extras', () => {
       const ys: unknown = [4, 5, 6, 7] as const;
 
       assert.deepStrictEqual(xs.fill(ys), [4, 5, 6]);
     });
 
-    test('keeps shorter arrays as-is and fills invalid entries', () => {
+    test('fills missing or invalid entries', () => {
       const ys: unknown = [4, '5'] as const;
 
-      assert.deepStrictEqual(xs.fill(ys), [4, 0]);
-    });
-
-    test('keeps the empty array', () => {
-      const ys: unknown = [] as const;
-
-      assert.deepStrictEqual(xs.fill(ys), []);
+      assert.deepStrictEqual(xs.fill(ys), [4, 0, 0]);
     });
 
     test('fill with the default value for non-array input', () => {
@@ -207,27 +220,27 @@ describe(arrayAtMostLength, () => {
   describe('size outside SmallUint', () => {
     // A literal beyond `SmallUint`'s range selects the fallback overload, whose
     // result length is unconstrained.
-    const big = arrayAtMostLength(100, number());
+    const big = minLengthTuple(100, number());
 
     expectType<TypeOf<typeof big>, readonly number[]>('=');
 
     expectType<typeof big.defaultValue, readonly number[]>('=');
 
     // A plain `number` (not a literal) also selects the fallback overload.
-    const dynamicSize: number = 2;
+    const dynamicSize: number = 5;
 
-    const dyn = arrayAtMostLength(dynamicSize, number());
+    const dyn = minLengthTuple(dynamicSize, number());
 
     expectType<TypeOf<typeof dyn>, readonly number[]>('=');
 
-    test('default value is the empty array', () => {
-      assert.deepStrictEqual(big.defaultValue, []);
+    test('default value has the requested length', () => {
+      expect(big.defaultValue).toHaveLength(100);
     });
 
-    test('validates the maximum length at runtime', () => {
-      assert.isTrue(dyn.is([0, 0]));
+    test('validates the minimum length at runtime', () => {
+      assert.isTrue(dyn.is([0, 0, 0, 0, 0]));
 
-      assert.isFalse(dyn.is([0, 0, 0]));
+      assert.isFalse(dyn.is([0, 0, 0, 0]));
     });
   });
 });
