+++ src/primitives/number.mts
@@ -40,7 +40,7 @@ export function number<C extends Constraints>(
 
   if (
     constraints === undefined ||
-    Arr.isArrayOfLength(Object.keys(constraints), 0)
+    Arr.isFixedLengthTuple(Object.keys(constraints), 0)
   ) {
     return baseType;
   }
