+++ src/primitives/string.mts
@@ -29,7 +29,7 @@ export function string<C extends RawConstraints>(
 
   if (
     constraints === undefined ||
-    Arr.isArrayOfLength(Object.keys(constraints), 0)
+    Arr.isFixedLengthTuple(Object.keys(constraints), 0)
   ) {
     return baseType;
   }
